+++ Common/src/main/java/com/tristankechlo/wool_collection/init/ModRegistry.java
@@ -1,12 +1,12 @@
 package com.tristankechlo.wool_collection.init;
 
 import com.tristankechlo.wool_collection.TheWoolCollection;
-import com.tristankechlo.wool_collection.blocks.WoolProcessorBlock;
-import com.tristankechlo.wool_collection.container.WoolProcessorContainer;
+import com.tristankechlo.wool_collection.blocks.WeavingStationBlock;
+import com.tristankechlo.wool_collection.container.WeavingStationContainer;
 import com.tristankechlo.wool_collection.platform.RegistrationProvider;
 import com.tristankechlo.wool_collection.platform.RegistryObject;
 import com.tristankechlo.wool_collection.platform.Services;
-import com.tristankechlo.wool_collection.recipe.WoolProcessorRecipe;
+import com.tristankechlo.wool_collection.recipe.WeavingStationRecipe;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.BlockItem;
@@ -27,10 +27,10 @@ public final class ModRegistry {
     public static final RegistrationProvider<RecipeSerializer<?>> RECIPE_SERIALIZERS = RegistrationProvider.get(BuiltInRegistries.RECIPE_SERIALIZER, TheWoolCollection.MOD_ID);
 
     /* Content */
-    public static final RegistryObject<Block> WOOL_PROCESSOR_BLOCK = BLOCKS.register("wool_processor", WoolProcessorBlock::new);
-    public static final RegistryObject<BlockItem> WOOL_PROCESSOR_ITEM = ITEMS.register("wool_processor", () -> new BlockItem(WOOL_PROCESSOR_BLOCK.get(), new Item.Properties()));
-    public static final RegistryObject<MenuType<WoolProcessorContainer>> WOOL_PROCESSOR_CONTAINER = CONTAINERS.register("wool_processor", Services.PLATFORM.buildContainer());
-    public static final RegistryObject<RecipeType<WoolProcessorRecipe>> WOOL_PROCESSOR_RECIPE_TYPE = RECIPES.register("wool_processor", () -> new RecipeType<>() {});
-    public static final RegistryObject<RecipeSerializer<WoolProcessorRecipe>> WOOL_PROCESSOR_RECIPE_SERIALIZER = RECIPE_SERIALIZERS.register("wool_processor", WoolProcessorRecipe.Serializer::new);
+    public static final RegistryObject<Block> WEAVING_STATION_BLOCK = BLOCKS.register("weaving_station", WeavingStationBlock::new);
+    public static final RegistryObject<BlockItem> WEAVING_STATION_ITEM = ITEMS.register("weaving_station", () -> new BlockItem(WEAVING_STATION_BLOCK.get(), new Item.Properties()));
+    public static final RegistryObject<MenuType<WeavingStationContainer>> WEAVING_STATION_CONTAINER = CONTAINERS.register("weaving_station", Services.PLATFORM.buildContainer());
+    public static final RegistryObject<RecipeType<WeavingStationRecipe>> WEAVING_STATION_RECIPE_TYPE = RECIPES.register("weaving_station", () -> new RecipeType<>() {});
+    public static final RegistryObject<RecipeSerializer<WeavingStationRecipe>> WEAVING_STATION_RECIPE_SERIALIZER = RECIPE_SERIALIZERS.register("weaving_station", WeavingStationRecipe.Serializer::new);
 
 }