+++ Common/src/main/java/com/tristankechlo/wool_collection/jei/WeavingStationRecipeCategory.java
@@ -1,9 +1,9 @@
 package com.tristankechlo.wool_collection.jei;
 
 import com.tristankechlo.wool_collection.TheWoolCollection;
-import com.tristankechlo.wool_collection.blocks.WoolProcessorBlock;
+import com.tristankechlo.wool_collection.blocks.WeavingStationBlock;
 import com.tristankechlo.wool_collection.init.ModRegistry;
-import com.tristankechlo.wool_collection.recipe.WoolProcessorRecipe;
+import com.tristankechlo.wool_collection.recipe.WeavingStationRecipe;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
 import mezz.jei.api.helpers.IGuiHelper;
@@ -18,21 +18,21 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 
-public class WoolProcessorRecipeCategory implements IRecipeCategory<WoolProcessorRecipe> {
+public class WeavingStationRecipeCategory implements IRecipeCategory<WeavingStationRecipe> {
 
-    private static final ResourceLocation LOCATION = new ResourceLocation(TheWoolCollection.MOD_ID, "textures/gui/wool_processor.png");
+    private static final ResourceLocation LOCATION = new ResourceLocation(TheWoolCollection.MOD_ID, "textures/gui/weaving_station.png");
     private final IDrawable background;
     private final IDrawable icon;
     private final Component localizedName;
 
-    public WoolProcessorRecipeCategory(IGuiHelper guiHelper) {
+    public WeavingStationRecipeCategory(IGuiHelper guiHelper) {
         background = guiHelper.createDrawable(LOCATION, 0, 213, 90, 43);
-        localizedName = WoolProcessorBlock.getContainerName();
-        icon = guiHelper.createDrawableItemStack(ModRegistry.WOOL_PROCESSOR_ITEM.get().getDefaultInstance());
+        localizedName = WeavingStationBlock.getContainerName();
+        icon = guiHelper.createDrawableItemStack(ModRegistry.WEAVING_STATION_ITEM.get().getDefaultInstance());
     }
 
     @Override
-    public RecipeType<WoolProcessorRecipe> getRecipeType() {
+    public RecipeType<WeavingStationRecipe> getRecipeType() {
         return JustEnoughItemsPlugin.RECIPE_TYPE;
     }
 
@@ -52,18 +52,18 @@ public class WoolProcessorRecipeCategory implements IRecipeCategory<WoolProcesso
     }
 
     @Override
-    public void setRecipe(IRecipeLayoutBuilder builder, WoolProcessorRecipe recipe, IFocusGroup focuses) {
+    public void setRecipe(IRecipeLayoutBuilder builder, WeavingStationRecipe recipe, IFocusGroup focuses) {
         builder.addSlot(RecipeIngredientRole.INPUT, 3, 3).addIngredients(recipe.getInputTop());
         builder.addSlot(RecipeIngredientRole.INPUT, 3, 24).addIngredients(recipe.getInputBottom());
         builder.addSlot(RecipeIngredientRole.OUTPUT, 67, 14).addItemStack(getResultItem(recipe));
     }
 
     @Override
-    public boolean isHandled(WoolProcessorRecipe recipe) {
+    public boolean isHandled(WeavingStationRecipe recipe) {
         return !recipe.isSpecial();
     }
 
-    private static ItemStack getResultItem(WoolProcessorRecipe recipe) {
+    private static ItemStack getResultItem(WeavingStationRecipe recipe) {
         Minecraft minecraft = Minecraft.getInstance();
         ClientLevel level = minecraft.level;
         if (level == null) {