+++ generator/mc/fence_gate.py
@@ -6,7 +6,7 @@ from mc.template import RecipeTemplate, RecoloringRecipeTemplate, Template, Empt
 
 """
 => Recipe
-can be crafted inside the wool_processor and can be recolored from other wool fence gates
+can be crafted inside the weaving_station and can be recolored from other wool fence gates
 
 => Advancement
 white fence has a separate advancement, triggerd by crafting white wool