+++ Changelog.md
@@ -3,7 +3,7 @@
 ### Version 1.20.1 - 1.1
 
 - fix fence_gates and buttons not making sounds sometimes
-- add new block `wool_processor`
+- add new block `weaving_station`
     - can be used to convert wool blocks to the wool blocks added by this mod
     - all previous recipes are no longer available
     - fixes recipe conflicts with some vanilla recipes